pyproject.toml
[tool.poetry]
name = "idiomapp"
version = "0.1.0"
description = "Interactive graph visualization application"
authors = ["sg"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
networkx = "^3.4.0"
matplotlib = "^3.10.0"
python-dotenv = "^1.1.0"
streamlit = "^1.44.0"
pyvis = "^0.3.2"
ollama = "^0.4.7"
gtts = "^2.5.4"
langdetect = "^1.0.9"
jinja2 = "^3.1.6"  # For FastAPI templates

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
watchfiles = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
api = "idiomapp.api.app:start"
graph = "idiomapp.streamlit.app:main" 